# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Job scheduling and task management service for Minisource",
        "title": "Scheduler Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5003",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/executions": {
            "get": {
                "description": "List executions with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "executions"
                ],
                "summary": "List executions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by job ID",
                        "name": "job_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by start time (RFC3339)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by end time (RFC3339)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.JobExecution"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/executions/stats": {
            "get": {
                "description": "Get statistics about executions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "executions"
                ],
                "summary": "Get execution statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start time (RFC3339)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/executions/{id}": {
            "get": {
                "description": "Get an execution by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "executions"
                ],
                "summary": "Get an execution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.JobExecution"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/executions/{id}/cancel": {
            "post": {
                "description": "Cancel a pending or running execution",
                "tags": [
                    "executions"
                ],
                "summary": "Cancel an execution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/jobs/{job_id}/executions": {
            "get": {
                "description": "List executions for a specific job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "executions"
                ],
                "summary": "List executions by job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.JobExecution"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_minisource_scheduler_internal_models.ExecutionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "retrying",
                "cancelled",
                "timeout"
            ],
            "x-enum-varnames": [
                "ExecutionStatusPending",
                "ExecutionStatusRunning",
                "ExecutionStatusCompleted",
                "ExecutionStatusFailed",
                "ExecutionStatusRetrying",
                "ExecutionStatusCancelled",
                "ExecutionStatusTimeout"
            ]
        },
...
        "github_com_minisource_scheduler_internal_models.JobExecution": {
            "type": "object",
            "properties": {
                "attempt": {
                    "description": "Current attempt number",
                    "type": "integer"
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration_ms": {
                    "description": "Duration in milliseconds",
                    "type": "integer"
                },
                "error": {
                    "description": "Error message",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "request": {
                    "description": "Request sent",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "response": {
                    "description": "Response received",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "scheduled_at": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_minisource_scheduler_internal_models.ExecutionStatus"
                },
                "status_code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "tenant_id": {
                    "type": "string"
                },
                "trace_id": {
                    "description": "Distributed trace ID",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "worker_id": {
                    "description": "ID of worker executing",
                    "type": "string"
                }
            }
        },
...
        "internal_handler.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_handler.Meta": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "internal_handler.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_handler.ErrorInfo"
                },
                "meta": {
                    "$ref": "#/definitions/internal_handler.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}